+++ gcmf_editor.py
@@ -1,5 +1,6 @@
 import bpy
 from .gcmf_node import GCMFTextureNode, collect_gcmf_texture_nodes
+from .ui import draw_checkbox_row, draw_checkbox_column
 
 NAME_ID_PROPERTY_VIEWER_BASE = 'OBJECT_PT_GCMF_{0}_Viewer'
 NAME_ID_PROPERTY_EDITOR_BASE = 'OBJECT_PT_GCMF_{0}_Editor'
@@ -39,79 +40,7 @@ MSG_HEX_8 = '{0}: {1:08X}'
 MSG_LABEL_SHOW = '{0}: {1}'
 MSG_LABEL_EDIT = '{0}:'
 
-def draw_checkbox_row(box: bpy.types.UILayout, data: bpy.types.AnyType, show_propertys: list[bool], show_property_keys: list[str], label_texts: list[str], data_masks: list[bool], property :str, label_text: str):
-    """Convert a list of boolean flags to a column of checkboxes with labels."""
-    idx = show_property_keys.index(property)
-    show_property = show_propertys[idx]
-    box.prop(data, 'show_propertys', index=idx, text=label_text, icon='TRIA_DOWN' if show_property else 'TRIA_RIGHT', emboss=False)
-    if show_property:
-        _checkbox = box.box()
-        for i, _label_text in enumerate(label_texts):
-            if not data_masks[i]:
-                continue
-            row = _checkbox.row()
-            row.prop(data, property, index=i, text=_label_text)
 
-def draw_checkbox_column(box: bpy.types.UILayout, data: bpy.types.AnyType, show_propertys: list[bool], show_property_keys: list[str], flags: list[bool], property :str, label_text: str):
-    idx = show_property_keys.index(property)
-    show_property = show_propertys[idx]
-    box.prop(data, 'show_propertys', index=idx, text=label_text, icon='TRIA_DOWN' if show_property else 'TRIA_RIGHT', emboss=False)
-    if show_property:
-        _checkbox = box.box()
-        length=len(flags)
-        col = _checkbox.column_flow(columns=length)
-        for i in range(length):
-            col.prop(data, property, index=i, text='')
-
-# # GCMF_Texture Add Button
-# # GCMFTextureNode をノードツリーに追加する
-# class GCMF_OT_Add_GCMFTexture(bpy.types.Operator):
-#     bl_idname = "gcmf.add_texture"
-#     bl_label = "Add GCMF Texture"
-#     bl_description = "Add a new GCMFTextureNode to the active material's node tree (max 3)"
-
-#     def execute(self, context):
-#         mat = context.active_object.active_material
-#         if mat is None or not mat.use_nodes:
-#             self.report({'WARNING'}, "No active material with nodes.")
-#             return {'CANCELLED'}
-
-#         existing = collect_gcmf_texture_nodes(mat)
-#         if len(existing) >= 3:
-#             self.report({'WARNING'}, "Maximum of 3 GCMF Texture nodes allowed per material.")
-#             return {'CANCELLED'}
-
-#         nodes = mat.node_tree.nodes
-#         new_node = nodes.new('GCMFTextureNode')
-#         new_node.label = f'GCMF Tex [{len(existing)}]'
-#         # 既存ノードの下に配置
-#         new_node.location = (-400, 300 - len(existing) * 320)
-#         return {'FINISHED'}
-
-
-# # GCMF_Texture Remove Button
-# # ノード名（bl_label + 連番）ではなく order_index を使ってノードを特定する
-# class GCMF_OT_Remove_GCMFTexture(bpy.types.Operator):
-#     bl_idname = "gcmf.remove_texture"
-#     bl_label = "Delete GCMF Texture"
-#     bl_description = "Remove the GCMFTextureNode at the specified slot from the node tree"
-
-#     # collect_gcmf_texture_nodes() の結果リスト上のインデックス
-#     index: bpy.props.IntProperty()
-
-#     def execute(self, context):
-#         mat = context.active_object.active_material
-#         if mat is None or not mat.use_nodes:
-#             return {'CANCELLED'}
-
-#         gcmf_nodes = collect_gcmf_texture_nodes(mat)
-#         if self.index >= len(gcmf_nodes):
-#             self.report({'WARNING'}, f"GCMF Texture [{self.index}] not found.")
-#             return {'CANCELLED'}
-
-#         mat.node_tree.nodes.remove(gcmf_nodes[self.index])
-#         return {'FINISHED'}
-    
 # GCMF Object Setting Show Panel
 class OBJECT_PT_GCMF_Object_Viewer(bpy.types.Panel):
     bl_idname = NAME_ID_PROPERTY_VIEWER_BASE.format('Object')
@@ -136,6 +65,7 @@ class OBJECT_PT_GCMF_Object_Viewer(bpy.types.Panel):
         except:
             pass # do nothing
 
+
 # GCMF Object Setting Edit Panel
 class OBJECT_PT_GCMF_Object_Editor(bpy.types.Panel):
     bl_parent_id = NAME_ID_PROPERTY_VIEWER_BASE.format('Object')
@@ -154,12 +84,12 @@ class OBJECT_PT_GCMF_Object_Editor(bpy.types.Panel):
             # attribute
             self.layout.prop(gcmf_object, "attribute", text="attribute")
 
-            self.layout.prop(gcmf_object, "transparent_material_count", text="Transparent Material Count")
+            self.layout.prop(gcmf_object, "transparent_count", text="Transparent Material Count")
 
         # If faild to get "gcmf_object" from active Object
         except Exception as e:
             self.layout.label(text=f"Error occurred: {e}")
-#            pass #do nothing
+
 
 # GCMF Material Setting Show Panel
 class OBJECT_PT_GCMF_Material_Viewer(bpy.types.Panel):
@@ -177,6 +107,9 @@ class OBJECT_PT_GCMF_Material_Viewer(bpy.types.Panel):
         return val
 
     def draw(self, context):
+        if not hasattr(bpy.context.active_object.active_material, 'gcmf_material'):
+            self.layout.label(text=MSG_NOT_FOUND_MATERIAL)
+            return
         try:
             gcmf_material = bpy.context.active_object.active_material.gcmf_material
             show_property_keys = [
@@ -236,13 +169,11 @@ class OBJECT_PT_GCMF_Material_Viewer(bpy.types.Panel):
             for i, _unk0x40 in enumerate(gcmf_material.unk0x40):
                 val += (int(_unk0x40) << (31 - i))
             self.layout.label(text=MSG_HEX_8.format(NAME_GXMDLVIEW_OBJ_UNK40, val))
-
-            # self.layout.operator("gcmf.add_texture", icon='ADD')
             
         # If faild to get "gcmf_material" from active Material
         except Exception as e:
             self.layout.label(text=f"Error occurred: {e}")
-#            self.layout.label(text=MSG_NOT_FOUND_MATERIAL)
+
 
 # GCMF Material Setting Edit Panel
 class OBJECT_PT_GCMF_Material_Editor(bpy.types.Panel):
@@ -254,6 +185,9 @@ class OBJECT_PT_GCMF_Material_Editor(bpy.types.Panel):
     bl_context = "material"
 
     def draw(self, context):
+        if not hasattr(bpy.context.active_object.active_material, 'gcmf_material'):
+            self.layout.label(text=MSG_NOT_FOUND_MATERIAL)
+            return
         try:
             gcmf_material = bpy.context.active_object.active_material.gcmf_material
             show_property_keys = [
@@ -284,7 +218,6 @@ class OBJECT_PT_GCMF_Material_Editor(bpy.types.Panel):
 
 
 # GCMF Texture Setting Show Panel
-# bpy.types.Texture is gone in Blender 4.x.
 # These panels now show texture settings from the active material's gcmf_textures collection.
 
 class OBJECT_PT_GCMF_Texture_Viewer(bpy.types.Panel):
@@ -295,13 +228,15 @@ class OBJECT_PT_GCMF_Texture_Viewer(bpy.types.Panel):
     bl_context = "material"
 
     def draw(self, context):
+        mat = bpy.context.active_object.active_material
+        if not hasattr(mat, 'use_nodes'):
+            self.layout.label(text=MSG_NOT_FOUND_MATERIAL)
+            return
         try:
-            mat = bpy.context.active_object.active_material
             gcmf_nodes = collect_gcmf_texture_nodes(mat)
 
             if len(gcmf_nodes) == 0:
                 self.layout.label(text=MSG_NOT_FOUND_TEXTURE)
-                # self.layout.operator("gcmf.add_texture", icon='ADD')
                 return
 
             for idx, gcmf_node in enumerate(gcmf_nodes):
@@ -337,101 +272,5 @@ class OBJECT_PT_GCMF_Texture_Viewer(bpy.types.Panel):
                     val = sum(int(b) << (31 - i) for i, b in enumerate(gcmf_node.unk0x10))
                     box.label(text=MSG_HEX_8.format(NAME_GXMDLVIEW_TEX_UNK10, val))
 
-            #     remove_button = self.layout.operator("gcmf.remove_texture", icon='REMOVE', text="")
-            #     remove_button.index = idx
-
-            # self.layout.operator("gcmf.add_texture", icon='ADD')
-
         except Exception as e:
             self.layout.label(text=f"Error occurred: {e}")
-
-
-# class OBJECT_PT_GCMF_Texture_Editor(bpy.types.Panel):
-#     bl_parent_id = NAME_ID_PROPERTY_VIEWER_BASE.format('Texture')
-#     bl_idname = NAME_ID_PROPERTY_EDITOR_BASE.format('Texture')
-#     bl_label = NAME_LABEL_PROPERTY_BASE.format('Texture Editor')
-#     bl_space_type = 'PROPERTIES'
-#     bl_region_type = 'WINDOW'
-#     bl_context = "material"
-
-#     def draw(self, context):
-#         try:
-#             mat = bpy.context.active_object.active_material
-#             gcmf_texture_nodes = collect_gcmf_texture_nodes(mat)
-
-#             if len(gcmf_texture_nodes) == 0:
-#                 self.layout.label(text=MSG_NOT_FOUND_TEXTURE)
-#                 return
-
-#             show_property_keys = [
-#                 'unk0x00', 'mipmap', 'uv_wrap', 'anisotropy',
-#                 'unk0x0C', 'is_swappable', 'unk0x10'
-#             ]
-#             for idx, gcmf_texture in enumerate(gcmf_texture_nodes):
-#                 is_open = gcmf_texture.show_gcmf_textures_edit
-#                 self.layout.prop(gcmf_texture, "show_gcmf_textures_edit", index=1,
-#                                  icon='TRIA_DOWN' if is_open else 'TRIA_RIGHT',
-#                                  emboss=False,
-#                                  text=f"Edit GCMF Texture [{idx}]")
-#                 if is_open:
-#                     box = self.layout.box()
-#                     # image
-#                     box.template_ID(gcmf_texture, "image", open="image.open")
-#                     # texture_index
-#                     box.label(text=NAME_GXMDLVIEW_TEX_TPL_IDX)
-#                     box.prop(gcmf_texture, "texture_index", text="value")
-#                     # unk0x00
-#                     unk0x00_labels = [''] * len(gcmf_texture.unk0x00)
-#                     unk0x00_labels[10] = 'Enable carcmntex.tpl Texture (F-ZERO GX)'
-#                     unk0x00_labels[11] = 'Enable normal reflection'
-#                     unk0x00_labels_mask = [False] * len(gcmf_texture.unk0x00)
-#                     unk0x00_labels_mask[10] = True
-#                     unk0x00_labels_mask[11] = True
-#                     draw_checkbox_row(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                      label_texts=unk0x00_labels, data_masks=unk0x00_labels_mask,
-#                                      property='unk0x00', label_text='unk0x00')
-#                     # mipmap
-#                     mip_map_labels = [
-#                         'unknown7', 'unknown6', 'unknown5', 'unknown4',
-#                         'near', 'unknown2', 'unknown1', 'enable'
-#                     ]
-#                     draw_checkbox_row(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                      label_texts=mip_map_labels,
-#                                      data_masks=[True] * len(gcmf_texture.mipmap),
-#                                      property='mipmap', label_text='MIPMAP')
-#                     # uv_wrap
-#                     uv_wrap_labels = [
-#                         'unknown7', 'unknown6', 'Y-Mirror', 'Y-Repeat',
-#                         'X-Mirror', 'X-Repeat', 'unknown1', 'unknown0'
-#                     ]
-#                     draw_checkbox_row(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                      label_texts=uv_wrap_labels,
-#                                      data_masks=[True] * len(gcmf_texture.uv_wrap),
-#                                      property='uv_wrap', label_text='UV WRAP')
-#                     # unk0x06
-#                     box.label(text=MSG_LABEL_EDIT.format(NAME_GXMDLVIEW_TEX_UNK06))
-#                     box.prop(gcmf_texture, "unk0x06", text="value")
-#                     # Anisotropy
-#                     anisotropy_labels = [
-#                         'unknown7', 'unknown6', 'unknown5', 'unknown4',
-#                         'unknown3', 'aniso4', 'aniso2', 'aniso1'
-#                     ]
-#                     draw_checkbox_row(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                      label_texts=anisotropy_labels,
-#                                      data_masks=[True] * len(gcmf_texture.anisotropy),
-#                                      property='anisotropy', label_text='Anisotropy')
-#                     # unk0x0C
-#                     draw_checkbox_column(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                         flags=gcmf_texture.unk0x0C,
-#                                         property='unk0x0C', label_text=NAME_GXMDLVIEW_TEX_UNK0C)
-#                     # is Swappable
-#                     draw_checkbox_column(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                         flags=gcmf_texture.is_swappable,
-#                                         property='is_swappable', label_text='is Swappable (0x0D)')
-#                     # unk0x10
-#                     draw_checkbox_column(box, gcmf_texture, gcmf_texture.show_propertys, show_property_keys,
-#                                         flags=gcmf_texture.unk0x10,
-#                                         property='unk0x10', label_text=NAME_GXMDLVIEW_TEX_UNK10)
-
-        # except Exception as e:
-        #     self.layout.label(text=f"Error occurred: {e}")
